requirements.txt
nose
sphinx
telethon
pyyaml